messenger_bot/requirements.txt
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
scikit-learn==0.20.0
scipy==1.1.0
sklearn==0.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
pymessenger
google-cloud-storage
